refactor(mobile): move the migration's style modules out of the route tree
expo-router registers every file under `app/` as a route, so the eleven style
modules this migration extracted there each produced
`Route "…" is missing the required default export` and became a navigable path
that renders nothing. Measured on a device: 19 such warnings before, 8 after.

Destinations follow the existing `src/onboarding/mobile-onboarding-styles.ts`
precedent — feature-area directory under `src/`, route stays in `app/`:
voice-settings-styles to `src/dictation/` (it styles the screen that configures
dictation), the three host-screen modules to a new `src/host/`, the seven
tasks-screen modules to the existing `src/tasks/`. Only `src/host/` is new;
every other feature area already has its directory, host overview was the one
that still kept all its code inside `app/`.

Moved with `git mv` so the rename is recorded. No behaviour change — only import
paths move (the modules' own `mobile-theme` depth, the three screen consumers,
the two merge modules' sibling imports, and eight dynamic imports in the factory
suite).

The remaining eight warnings are pre-existing files no batch owns, and are not
all stylesheets — one is types-only and one is a component. Left alone here.
