+++ mobile/app/voice-settings.tsx
@@ -5,7 +5,7 @@ import { useFocusEffect, useRouter } from 'expo-router'
 import { ChevronLeft, ChevronRight } from 'lucide-react-native'
 import { spacing } from '../src/theme/mobile-theme'
 import { useTheme, useThemedStyles } from '../src/theme/theme-context'
-import { createVoiceSettingsStyles } from './voice-settings-styles'
+import { createVoiceSettingsStyles } from '../src/dictation/voice-settings-styles'
 import { loadHosts } from '../src/transport/host-store'
 import type { HostProfile } from '../src/transport/types'
 import { useAllHostClients } from '../src/transport/client-context'
