+++ mobile/src/tasks/tasks-screen-list-styles.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native'
-import { radii, spacing, type ThemeColors } from '../../../src/theme/mobile-theme'
+import { radii, spacing, type ThemeColors } from '../theme/mobile-theme'
 
 // Born themed — split from tasks.tsx (no max-lines headroom).
 export const createTasksScreenListStyles = (colors: ThemeColors) =>
